Отредактирован Converter, исправлен bestSeries, исправлен printStatistic,отредактирован changeStepGoal

+++ src/Converter.java
@@ -2,12 +2,12 @@ public class Converter {
 
     double convertToKm(int steps){
         int kmToCm = 100000;
-        int stepInCm = 75;
+        double stepInCm = 75;
         return (stepInCm * steps) / kmToCm;
     }
 
     double convertStepsToKilocalories(int steps){
-        int stepInCalories = 50;
+        double stepInCalories = 50;
         int kiloСaloriesToCalories = 1000;
         return (stepInCalories * steps) / kiloСaloriesToCalories;
     }
